server/requirements.txt
asgiref==3.5.2
autopep8==1.6.0
Django==4.0.5
django-cors-headers==3.13.0
django-environ==0.9.0
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.0
djongo==1.3.6
dnspython==2.2.1
Pillow==9.1.1
pycodestyle==2.8.0
PyJWT==1.7.1
pymongo==3.12.3
pytz==2022.1
sqlparse==0.2.4
toml==0.10.2
gunicorn==20.1.0
